+++ src/request/user.rs
@@ -1,1 +1,2 @@
+pub mod user_one;
 pub mod users;
