+++ src/response/user/user_one.rs
@@ -1,9 +1,8 @@
+use crate::record::user::ReadOnlyUserRecord;
 use serde::{Deserialize, Serialize};
-
-use crate::response::user::UserRecord;
 
 #[derive(Debug, Deserialize, Clone, Serialize)]
 pub struct GetUserResponse {
     pub status: u16,
-    pub body: UserRecord,
+    pub body: ReadOnlyUserRecord,
 }
